apps/authdemo/serializers.py
from rest_framework import serializers
from .models import User

class UserSerializers(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('uuid','username','telephone','is_active','is_staff')
