requirements.txt
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-dysmsapi==1.0.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-debug-toolbar==1.10.1
django-shortuuidfield==0.1.3
djangorestframework==3.8.2
idna==2.6
mysqlclient==1.3.12
Pillow==5.2.0
PyMySQL==0.8.1
python-memcached==1.59
pytz==2018.4
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
